convert group outputs to normal p2pkh or p2sh.
This is a failed attempt at c++, is at best pseudo code, and doesn't
compile. The goal is to detect when a scriptPubKey includes OP_GROUP,
then detect if the script is p2pkh or p2sh, then remove the group parts
so hash is calculated from normal output.

While making this update I noticed that in Mempool.cpp there is only a
loop around outputs to compute scriptHashes.  Why not also loop
through inputs?  Clients may be subscribed to a p2sh input that has no
relevant output, would the client receive a notification?